+++ resources/views/pagination/livewire-default.blade.php
@@ -4,14 +4,14 @@
             <div class="flex flex-1 justify-between sm:hidden">
                 <span>
                     @if ($paginator->onFirstPage())
-                        <x-chief-table-new::button
+                        <x-chief-table::button
                             size="sm"
                             color="white"
                             iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M15 6C15 6 9.00001 10.4189 9 12C8.99999 13.5812 15 18 15 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
                             class="pointer-events-none text-grey-400"
                         >
                             Vorige
-                        </x-chief-table-new::button>
+                        </x-chief-table::button>
                     @else
                         <button
                             type="button"
@@ -19,13 +19,13 @@
                             wire:loading.attr="disabled"
                             dusk="previousPage{{ $paginator->getPageName() == 'page' ? '' : '.' . $paginator->getPageName() }}.before"
                         >
-                            <x-chief-table-new::button
+                            <x-chief-table::button
                                 size="sm"
                                 color="white"
                                 iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M15 6C15 6 9.00001 10.4189 9 12C8.99999 13.5812 15 18 15 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
                             >
                                 Vorige
-                            </x-chief-table-new::button>
+                            </x-chief-table::button>
                         </button>
                     @endif
                 </span>
@@ -38,23 +38,23 @@
                             wire:loading.attr="disabled"
                             dusk="nextPage{{ $paginator->getPageName() == 'page' ? '' : '.' . $paginator->getPageName() }}.before"
                         >
-                            <x-chief-table-new::button
+                            <x-chief-table::button
                                 size="sm"
                                 color="white"
                                 iconRight='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M9.00005 6C9.00005 6 15 10.4189 15 12C15 13.5812 9 18 9 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
                             >
                                 Volgende
-                            </x-chief-table-new::button>
+                            </x-chief-table::button>
                         </button>
                     @else
-                        <x-chief-table-new::button
+                        <x-chief-table::button
                             size="sm"
                             color="white"
                             iconRight='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M9.00005 6C9.00005 6 15 10.4189 15 12C15 13.5812 9 18 9 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
                             class="pointer-events-none text-grey-400"
                         >
                             Volgende
-                        </x-chief-table-new::button>
+                        </x-chief-table::button>
                     @endif
                 </span>
             </div>
@@ -77,7 +77,7 @@
                             {{-- Previous Page Link --}}
                             @if ($paginator->onFirstPage())
                                 <span aria-disabled="true" aria-label="{{ __('pagination.previous') }}">
-                                    <x-chief-table-new::button
+                                    <x-chief-table::button
                                         size="sm"
                                         color="white"
                                         iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M15 6C15 6 9.00001 10.4189 9 12C8.99999 13.5812 15 18 15 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
@@ -92,7 +92,7 @@
                                     rel="prev"
                                     aria-label="{{ __('pagination.previous') }}"
                                 >
-                                    <x-chief-table-new::button
+                                    <x-chief-table::button
                                         size="sm"
                                         color="white"
                                         iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M15 6C15 6 9.00001 10.4189 9 12C8.99999 13.5812 15 18 15 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
@@ -107,13 +107,13 @@
                             {{-- "Three Dots" Separator --}}
                             @if (is_string($element))
                                 <span aria-disabled="true">
-                                    <x-chief-table-new::button
+                                    <x-chief-table::button
                                         size="sm"
                                         color="white"
                                         class="pointer-events-none min-w-[1.875rem] rounded-l-none rounded-r-none font-normal text-grey-400 hover:relative"
                                     >
                                         {{ $element }}
-                                    </x-chief-table-new::button>
+                                    </x-chief-table::button>
                                 </span>
                             @endif
 
@@ -123,13 +123,13 @@
                                     <span wire:key="paginator-{{ $paginator->getPageName() }}-page{{ $page }}">
                                         @if ($page == $paginator->currentPage())
                                             <span aria-current="page">
-                                                <x-chief-table-new::button
+                                                <x-chief-table::button
                                                     size="sm"
                                                     color="white"
                                                     class="pointer-events-none min-w-[1.875rem] rounded-l-none rounded-r-none font-normal text-grey-400 hover:relative"
                                                 >
                                                     {{ $page }}
-                                                </x-chief-table-new::button>
+                                                </x-chief-table::button>
                                             </span>
                                         @else
                                             <button
@@ -137,13 +137,13 @@
                                                 wire:click="gotoPage({{ $page }}, '{{ $paginator->getPageName() }}')"
                                                 aria-label="{{ __('Go to page :page', ['page' => $page]) }}"
                                             >
-                                                <x-chief-table-new::button
+                                                <x-chief-table::button
                                                     size="sm"
                                                     color="white"
                                                     class="min-w-[1.875rem] rounded-l-none rounded-r-none font-normal hover:relative"
                                                 >
                                                     {{ $page }}
-                                                </x-chief-table-new::button>
+                                                </x-chief-table::button>
                                             </button>
                                         @endif
                                     </span>
@@ -161,7 +161,7 @@
                                     rel="next"
                                     aria-label="{{ __('pagination.next') }}"
                                 >
-                                    <x-chief-table-new::button
+                                    <x-chief-table::button
                                         size="sm"
                                         color="white"
                                         iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M9.00005 6C9.00005 6 15 10.4189 15 12C15 13.5812 9 18 9 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'
@@ -170,7 +170,7 @@
                                 </button>
                             @else
                                 <span aria-disabled="true" aria-label="{{ __('pagination.next') }}">
-                                    <x-chief-table-new::button
+                                    <x-chief-table::button
                                         size="sm"
                                         color="white"
                                         iconLeft='<svg viewBox="0 0 24 24" color="currentColor" fill="none"> <path d="M9.00005 6C9.00005 6 15 10.4189 15 12C15 13.5812 9 18 9 18" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" /> </svg>'