+++ src/Plugins/Tags/App/resources/views/tags/index.blade.php
@@ -3,7 +3,7 @@
         <x-chief::page.hero title="Tags" class="max-w-3xl"/>
     </x-slot>
 
-    <x-chief::page.grid class="max-w-3xl">
+    <x-chief::page.grid class="max-w-3xl mt-6">
         @foreach($tagGroups as $tagGroup)
             @php
                 $tagsByGroup = $tags->filter(function($tag) use ($tagGroup) {
@@ -11,7 +11,7 @@
                 });
             @endphp
 
-            <x-chief::window>
+            <div>
                 @if ($tagGroups->count() > 1)
                     <div id="taggroup-{{ $tagGroup->getTagGroupId() }}" class="flex items-start justify-between mb-2">
                         <div class="flex items-center gap-1">
@@ -53,33 +53,36 @@
 
                 @endif
 
-                <x-chief::table>
-                    <x-slot name="body">
+                <div class="border card-without-padding px-5 border-grey-200">
+                    <table class="min-w-full table-fixed divide-y divide-grey-200">
+                        <tbody class="divide-y divide-grey-200">
                         @forelse ($tagsByGroup as $tag)
-                            <x-chief::table.row>
-                                <x-chief::table.data>
-                                    <a
-                                        href="{{ route('chief.tags.edit', $tag->getTagId()) }}"
-                                        title="{{ $tag->getLabel() }}"
-                                        class="inline-flex items-center gap-x-3"
-                                    >
-                                        <svg class="w-2.5 h-2.5" style="fill: {{ $tag->getColor() }};" viewBox="0 0 6 6" aria-hidden="true">
-                                            <circle cx="3" cy="3" r="3" />
-                                        </svg>
+                            <tr>
+                                <td class="py-1.5 pl-3 text-left">
+                                    <div class="flex min-h-6 items-center gap-1.5">
+                                        <a
+                                            href="{{ route('chief.tags.edit', $tag->getTagId()) }}"
+                                            title="{{ $tag->getLabel() }}"
+                                            class="inline-flex items-center gap-x-3"
+                                        >
+                                            <svg class="w-2.5 h-2.5" style="fill: {{ $tag->getColor() }};" viewBox="0 0 6 6" aria-hidden="true">
+                                                <circle cx="3" cy="3" r="3" />
+                                            </svg>
 
-                                        <span class="font-medium body-dark hover:underline">{{ $tag->getLabel() }}</span>
-                                    </a>
-                                </x-chief::table.data>
-
-                                <x-chief::table.data>
-                                    @if($tag->getUsages() > 0)
-                                        <p>In gebruik op {{ $tag->getUsages() }} pagina's</p>
-                                    @else
-                                        <p>Niet in gebruik</p>
-                                    @endif
-                                </x-chief::table.data>
-
-                                <x-chief::table.data>
+                                            <span class="font-medium body-dark hover:underline">{{ $tag->getLabel() }}</span>
+                                        </a>
+                                    </div>
+                                </td>
+                                <td class="py-1.5 pl-3 text-left">
+                                    <div class="flex min-h-6 items-center gap-1.5">
+                                        @if($tag->getUsages() > 0)
+                                            <p>In gebruik op {{ $tag->getUsages() }} pagina's</p>
+                                        @else
+                                            <p>Niet in gebruik</p>
+                                        @endif
+                                    </div>
+                                </td>
+                                <td class="py-1.5 pr-3 text-right">
                                     <div class="flex justify-end -mr-2">
                                         <a href="{{ route('chief.tags.edit', $tag->getTagId()) }}">
                                             <x-chief::icon-button icon="icon-edit" color="grey" class="bg-white shadow-none text-grey-500" />
@@ -87,20 +90,21 @@
 
                                         @include('chief-tags::tags.delete')
                                     </div>
-                                </x-chief::table.data>
-                            </x-chief::table.row>
+                                </td>
+                            </tr>
                         @empty
-                            <x-chief::table.row>
-                                <x-chief::table.data>
+                            <tr>
+                                <td class="py-1.5 pl-3 text-left">
                                     <div class="min-h-[30px] flex items-center">
                                         <span class="text-grey-400">Deze groep bevat nog geen tags</span>
                                     </div>
-                                </x-chief::table.data>
-                            </x-chief::table.row>
+                                </td>
+                            </tr>
                         @endforelse
-                    </x-slot>
-                </x-chief::table>
-            </x-chief::window>
+                        </tbody>
+                    </table>
+                </div>
+            </div>
         @endforeach
 
         @if (count($tags) > 0)