+++ src/Table/Actions/BulkAction.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Table\Actions;
+
+class BulkAction extends Action
+{
+    protected string $view = 'chief-table::actions.bulk-action';
+}