+++ src/Table/Actions/RowAction.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Table\Actions;
+
+class RowAction extends Action
+{
+    protected string $view = 'chief-table::actions.row-action';
+}