+++ src/Table/UI/views/filters/search.blade.php
@@ -1,0 +1,8 @@
+<x-chief-table::filter.text>
+    <input
+        type="text"
+        wire:model.live.debounce.500ms="filters.{{ $getKey() }}"
+        placeholder="{{ $getPlaceholder() }}"
+        class="w-48"
+    />
+</x-chief-table::filter.text>