+++ src/Table/resources/cells/text.blade.php
@@ -1,8 +1,0 @@
-<x-chief::table.data>
-    <span
-        {{ $attributes->merge($getCustomAttributes()) }}
-        title="{{ $getHint() }}"
-    >
-        {!! $getValue() !!}
-    </span>
-</x-chief::table.data>