+++ src/TableNew/Columns/Header.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Thinktomorrow\Chief\TableNew\Columns;
-
-class Header extends ColumnItem
-{
-    protected string $view = 'chief-table-new::columns.header';
-}