+++ src/TableNew/UI/views/actions/bulk-action.blade.php
@@ -1,9 +1,0 @@
-@if ($hasLink())
-    <a href="{{ $getLink() }}" title="{{ $getDescription() }}">
-        <x-chief-table-new::button size="xs">{{ $getLabel() }}</x-chief-table-new::button>
-    </a>
-@else
-    <button wire:click="applyAction('{{ $getKey() }}')" title="{{ $getDescription() }}">
-        <x-chief-table-new::button size="xs">{{ $getLabel() }}</x-chief-table-new::button>
-    </button>
-@endif