+++ src/TableNew/UI/views/actions/row-action.blade.php
@@ -1,9 +1,0 @@
-@if($hasLink())
-    <a href="{{ $getLink() }}" title="{{ $getDescription() }}">
-        <x-chief-table-new::button size="sm">{{ $getLabel() }}</x-chief-table-new::button>
-    </a>
-@else
-    <button wire:click="applyBulkAction('{{ $getKey() }}')" title="{{ $getDescription() }}">
-        <x-chief-table-new::button size="sm">{{ $getLabel() }}</x-chief-table-new::button>
-    </button>
-@endif